+++ src/components/feature/main/PostItem.tsx
@@ -4,8 +4,8 @@ import Bookmark from '../../common/Bookmark';
 import SkillList from '../../common/SkillList';
 import StudyTypeIcon from '../../common/StudyTypeIcon';
 import { useNavigate } from 'react-router-dom';
-import { getDateFomat } from '../../../utils/dateUtils';
-import { PostCardType } from '../../../types/commonTypes';
+import { getDateFomat } from '../../../utils';
+import { PostCardType } from '../../../types';
 import studyIcon from '../../../assets/study-icon.svg';
 import projectIcon from '../../../assets/project-icon.svg';
 import Views from '../../common/Views';
