+++ src/components/feature/mypage/tabs/ManageProfile.tsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import { useMutation, useQuery } from 'react-query';
 import { useNavigate } from 'react-router-dom';
-import { skills, positions, career } from '../../../../constants/data';
-import { UserType } from '../../../../types/commonTypes';
+import { skills, positions, career } from '../../../../constants';
+import { UserType } from '../../../../types';
 import api from '../../../../services/api';
 import Input from '../../../common/Input';
 import CircleIcon from '../../../common/CircleIcon';
