+++ src/components/feature/postDetail/DetailHeader.tsx
@@ -6,9 +6,9 @@ import CircleIcon from '../../common/CircleIcon';
 import Bookmark from '../../common/Bookmark';
 import Views from '../../common/Views';
 import { useNavigate } from 'react-router-dom';
-import { getDateFomat } from '../../../utils/dateUtils';
+import { getDateFomat } from '../../../utils';
 import { ChevronLeftIcon } from '@radix-ui/react-icons';
-import { PostDetailType } from '../../../types/commonTypes';
+import { PostDetailType } from '../../../types';
 import * as AlertDialog from '@radix-ui/react-alert-dialog';
 import * as Menubar from '@radix-ui/react-menubar';
 import AlertModal from '../../common/AlertModal';
